# specs/001-fleet-roster-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Pickup MaaS API - 차량 및 승객명단 관리
  description: |
    B2B 송영 서비스를 위한 RESTful API

    **Features**:
    - 001번: 기관 차량 및 승객명단 관리
    - 002번: 주간보호 8시간 케어 스케줄 검증

    **Authentication**: Bearer JWT token required for all endpoints

    **API Conventions**:
    - All timestamps in ISO 8601 format (UTC)
    - All responses include standard error format
    - Pagination uses offset-based strategy
    - Phone numbers stored without hyphens (01012345678)
    - Time values in HH:MM format (24-hour)

  version: 1.0.0
  contact:
    name: Pickup MaaS Team
    email: dev@pickup-maas.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api-staging.pickup-maas.com/v1
    description: Staging environment
  - url: https://api.pickup-maas.com/v1
    description: Production environment
...
security:
  - BearerAuth: []
paths:
...
  /institutions:
    get:
      tags: [Institutions]
      summary: 기관 목록 조회 (Admin only)
      description: 시스템 관리자용 - 모든 기관 목록 조회
      operationId: getAllInstitutions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: institutionTypeId
          in: query
          description: 기관 유형 필터
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 기관 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInstitutions'

    post:
      tags: [Institutions]
      summary: 기관 생성
      description: 신규 기관 등록 (사업자등록번호 중복 불가)
      operationId: createInstitution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstitutionDto'
      responses:
        '201':
          description: 기관 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Institution'
        '409':
          $ref: '#/components/responses/ConflictError'
  /institutions/{institutionId}:
    get:
      tags: [Institutions]
      summary: 기관 상세 조회
      operationId: getInstitution
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
      responses:
        '200':
          description: 기관 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Institution'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags: [Institutions]
      summary: 기관 정보 수정
      operationId: updateInstitution
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInstitutionDto'
      responses:
        '200':
          description: 기관 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Institution'

    delete:
      tags: [Institutions]
      summary: 기관 삭제
      description: 기관 삭제 (CASCADE - 모든 차량, 그룹, 승객 삭제됨)
      operationId: deleteInstitution
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
      responses:
        '204':
          description: 기관 삭제 성공

  # ============================================================================
  # Vehicle Endpoints (차량)
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  parameters:
    InstitutionIdParam:
      name: institutionId
      in: path
      required: true
      description: 기관 ID (UUID)
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: 페이지 번호 (1부터 시작)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: 페이지당 항목 수
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  schemas:
...
    InstitutionType:
      type: object
      required:
        - id
        - typeCode
        - typeName
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        typeCode:
          type: string
          enum: [DAYCARE, GENERAL]
          description: 기관 유형 코드
        typeName:
          type: string
          description: 기관 유형명 (주간보호, 일반)
        minimumCareTimeHours:
          type: integer
          nullable: true
          description: 최소 케어 시간 (주간보호는 8, 일반은 null)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Institution:
      type: object
      required:
        - id
        - businessRegistrationNo
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        businessRegistrationNo:
          type: string
          pattern: '^\d{10}$'
          description: 사업자등록번호 (10자리, 하이픈 제거)
          example: '1234567890'
        name:
          type: string
          maxLength: 100
          description: 기관명
          example: '서울 주간보호센터'
        institutionTypeId:
          type: string
          format: uuid
          nullable: true
        institutionType:
          $ref: '#/components/schemas/InstitutionType'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateInstitutionDto:
      type: object
      required:
        - businessRegistrationNo
        - name
      properties:
        businessRegistrationNo:
          type: string
          pattern: '^\d{10}$'
          example: '1234567890'
        name:
          type: string
          maxLength: 100
          example: '서울 주간보호센터'
        institutionTypeId:
          type: string
          format: uuid
    UpdateInstitutionDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        institutionTypeId:
          type: string
          format: uuid
    PaginatedInstitutions:
      type: object
      required:
        - data
        - total
        - page
        - limit
        - totalPages
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Institution'
        total:
          type: integer
          description: 전체 항목 수
        page:
          type: integer
          description: 현재 페이지
        limit:
          type: integer
          description: 페이지당 항목 수
        totalPages:
          type: integer
          description: 전체 페이지 수

    # ========================================================================
    # Vehicle Schemas
    # ========================================================================
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - message
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: '요청 처리 중 오류가 발생했습니다'
        error:
          type: string
          example: 'Bad Request'
        details:
          type: array
          items:
            type: string
          description: 상세 에러 메시지 목록
  responses:
    NotFoundError:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            message: '요청한 리소스를 찾을 수 없습니다'
            error: 'Not Found'
    ConflictError:
      description: 리소스 충돌 (중복된 값 등)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 409
            message: '이미 존재하는 값입니다'
            error: 'Conflict'
...